move calendar logic to a service (makes more sense pattern wise)
